feat: add lazy variables
